# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Universal Document Vector Search Service",
    "description": "Микросервис для векторного поиска по документам с поддержкой метаданных, категорий и версионирования",
    "version": "1.2.0"
  },
  "paths": {
    "/v1/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Загрузка документов",
        "description": "Загружает один или несколько текстовых документов в указанную коллекцию.\n\n**Параметры запроса:**\n- `documents` — массив документов, каждый с полями:\n  - `text` — текст документа (обязательно)\n  - `payload` — метаданные (source_format, author и т.д.)\n  - `category_path` — иерархия категорий, например `[\"Документация\", \"Инструкции\"]`\n  - `filename` — имя файла (для генерации `source_id`)\n  - `title` — заголовок документа (для генерации `source_id`)\n  - `version` — версия документа (по умолчанию 1)\n- `collection_name` — имя коллекции в Qdrant (обязательно)\n\n**Особенности:**\n- Автоматически вычисляет хеш документа\n- Проверяет дубликаты по `source_id` и `doc_hash`\n- Поддерживает иерархию категорий\n- Возвращает список созданных `ids` и общее количество загруженных чанков\n\nЕсли ни один документ не был изменён или добавлен, возвращается сообщение  \n`Нет новых или изменившихся документов для загрузки`.",
        "operationId": "upload_documents_v1_documents_upload_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_v1_documents_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/files/upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "summary": "Загрузка одного файла",
        "description": "Загружает один файл (PDF, текст и др.) в указанную коллекцию.\n\nПоддерживает версионность: если файл с таким же содержимым уже существует, он не перезаписывается.  \nЕсли содержимое изменилось — создаётся новая версия.\n\nЕсли `skip_if_exists=true` и файл уже есть, возвращается статус `skipped`.  \nЕсли `source_id` не указан, он генерируется на основе `category_path` и имени файла.\n\n**Особенности:**\n- Автоматически вычисляет SHA256-хеш\n- Проверяет дубликаты по `source_id` и `doc_hash`\n- Поддерживает иерархию категорий\n- Возвращает `source_id`, `version`, `status` (`created`/`updated`/`skipped`)",
        "operationId": "upload_file_v1_files_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_v1_files_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_documents_v1_documents_upload_post": {
        "properties": {
          "documents": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents",
            "description": "JSON-массив документов (строкой). Если не передан, можно отправить обычное JSON-тело запроса."
          },
          "collection_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Name",
            "description": "Имя коллекции. Если не передано, можно отправить обычное JSON-тело запроса."
          }
        },
        "type": "object",
        "title": "Body_upload_documents_v1_documents_upload_post"
      },
      "Body_upload_file_v1_files_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "source_format": {
            "type": "string",
            "title": "Source Format"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "category_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Path"
          },
          "skip_if_exists": {
            "type": "boolean",
            "title": "Skip If Exists",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "file",
          "collection_name",
          "source_format"
        ],
        "title": "Body_upload_file_v1_files_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}